OneViewModelからGithubRepositoryを切り分け
