+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/search/CustomAdapter.kt
@@ -1,0 +1,4 @@
+package jp.co.yumemi.android.code_check.ui.search
+
+class CustomAdapter {
+}
