+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/search/OneViewModel.kt
@@ -3,77 +3,26 @@
  */
 package jp.co.yumemi.android.code_check.ui.search
 
-import android.content.Context
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import io.ktor.client.HttpClient
-import io.ktor.client.call.receive
-import io.ktor.client.engine.android.Android
-import io.ktor.client.request.get
-import io.ktor.client.request.header
-import io.ktor.client.request.parameter
-import io.ktor.client.statement.HttpResponse
-import jp.co.yumemi.android.code_check.R
+import androidx.lifecycle.viewModelScope
 import jp.co.yumemi.android.code_check.TopActivity.Companion.lastSearchDate
 import jp.co.yumemi.android.code_check.data.model.item
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.async
-import kotlinx.coroutines.runBlocking
-import org.json.JSONObject
+import jp.co.yumemi.android.code_check.data.repository.GithubRepository
+import kotlinx.coroutines.launch
 import java.util.Date
 
 /**
  * TwoFragment で使う
  */
-class OneViewModel(
-    val context: Context,
-) : ViewModel() {
-    // 検索結果
-    fun searchResults(inputText: String): List<item> =
-        runBlocking {
-            val client = HttpClient(Android)
-
-            return@runBlocking GlobalScope.async {
-                val response: HttpResponse =
-                    client?.get("https://api.github.com/search/repositories") {
-                        header("Accept", "application/vnd.github.v3+json")
-                        parameter("q", inputText)
-                    }
-
-                val jsonBody = JSONObject(response.receive<String>())
-
-                val jsonItems = jsonBody.optJSONArray("items")!!
-
-                val items = mutableListOf<item>()
-
-                /**
-                 * アイテムの個数分ループする
-                 */
-                for (i in 0 until jsonItems.length()) {
-                    val jsonItem = jsonItems.optJSONObject(i)!!
-                    val name = jsonItem.optString("full_name")
-                    val ownerIconUrl = jsonItem.optJSONObject("owner")!!.optString("avatar_url")
-                    val language = jsonItem.optString("language")
-                    val stargazersCount = jsonItem.optLong("stargazers_count")
-                    val watchersCount = jsonItem.optLong("watchers_count")
-                    val forksCount = jsonItem.optLong("forks_conut")
-                    val openIssuesCount = jsonItem.optLong("open_issues_count")
-
-                    items.add(
-                        item(
-                            name = name,
-                            ownerIconUrl = ownerIconUrl,
-                            language = context.getString(R.string.written_language, language),
-                            stargazersCount = stargazersCount,
-                            watchersCount = watchersCount,
-                            forksCount = forksCount,
-                            openIssuesCount = openIssuesCount,
-                        ),
-                    )
-                }
-
-                lastSearchDate = Date()
-
-                return@async items.toList()
-            }.await()
+class OneViewModel(private val repository: GithubRepository) : ViewModel() {
+    fun searchResults(inputText: String): LiveData<List<item>> {
+        val result = MutableLiveData<List<item>>()
+        viewModelScope.launch {
+            result.value = repository.searchRepositories(inputText)
+            lastSearchDate = Date() // 検索日時を更新
         }
+        return result
+    }
 }
